Set up controlled component tests
This commit introduces changes that enforce a controlled component architecture. They specify:
  *An input container component. When the user changes the content in the field, a callback is called. This callback is passed into this component's props from its parent component.
  *State in the parent component which is set by the input field's callback.
  *Props in a results container component which is set by the parent's state.

An existing test specifies that when the results container component props are set, they are rendered in the UI. So, this entire test suite ensures that when a user types input into an input field, the final calculation result is indeed rendered in the UI via a controlled component architecture.

These tests cover the functionality of the integration test which I had written previously, and which did not enforce a controlled component architecture. So, this commit removes that integration test.