Implement controlled component architecture
When a user types input into an input field, a calculation result is rendered in the UI. This is achieved via an architecture where the input change triggers a callback that sets a parent component's state. This parent state sets props in another child component which renders a calculation result.