Add test for an input sanitization function
Whenever the form is updated, a function should be called that will sanitize the input data into a number regardless of the input data type. If a number is entered into the form, that number should be returned from the sanitization function. If the data cannot be parsed into a number (such as a string or an empty input), the sanitization function should return 0.