Implement input sanitization function
Whenever the form is updated, a function is called that sanitizes the input data into a number regardless of the input data type. If a number is entered into the form, that number is returned from the sanitization function. If the data cannot be parsed into a number (such as a string or an empty input), the sanitization function returns 0.