Fix syntax to simulate click events with Enzyme
Previously, I tried to simulate click events with ReactTestUtils, and it wasn't working. As of now, I am finding the DOM nodes using Enzyme, suggesting that Enzyme should be used to simulate click events. This commit applies changes so that Enzyme is both finding the DOM nodes and simulating click events on them, which is working.